[#157] Restructure github services under webhooks module
