+++ .env
@@ -32,10 +32,11 @@ AWS_REGION=""
 ROLLBAR_POST_SERVER_ITEM_TOKEN=""
 ROLLBAR_POST_CLIENT_ITEM_TOKEN=""
 
-GITLAB_CLIENT_ID=
-GITLAB_CLIENT_SECRET=
+GITLAB_APP_ID=
+GITLAB_APP_SECRET=
 GITLAB_REDIRECT_URI=
 GITLAB_API_ENDPOINT=
+GITLAB_WEBHOOK_SECRET=
 GITLAB_DEPLOYQA_BOT_ID=
 GITLAB_DEPLOYQA_BOT_TOKEN=
 GITLAB_ENABLED=
