+++ app/controllers/webhooks/github_controller.rb
@@ -6,14 +6,14 @@ module Webhooks
     skip_before_action :login_if_not
 
     def create
-      wrapped_request = Github::WebhookRequestWrapper.build(request)
+      wrapped_request = Github::Webhooks::RequestWrapper.build(request)
       WebhookLog.create!(body: wrapped_request.parse_body,
                          event: wrapped_request.github_event,
                          integration_type: ProjectsConstants::Providers::GITHUB)
 
       WebhookResolverJob.perform_later(wrapped_request.serialize,
-                                       Github::WebhookRequestWrapper.to_s,
-                                       Github::WebhookResolver.to_s)
+                                       Github::Webhooks::RequestWrapper.to_s,
+                                       Github::Webhooks::Resolver.to_s)
       head :ok
     end
   end
