--- app/services/github/webhook_processors/revoke_auth.rb
@@ -1,17 +1,0 @@
-# frozen_string_literal: true
-
-module Github
-  module WebhookProcessors
-    class RevokeAuth
-      def initialize(body)
-        @wrapped_body = Github::Events::Authorization.new(payload: body)
-      end
-
-      def call
-        user = ::User.find_by(auth_provider: ProjectsConstants::Providers::GITHUB, auth_uid: @wrapped_body.id)
-        user.project_user_roles.destroy_all
-        ReturnValue.ok
-      end
-    end
-  end
-end
