+++ app/services/github/webhooks/processors/delete_project.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module Github
+  module Webhooks
+    module Processors
+      class DeleteProject
+        def initialize(body)
+          @body = body
+        end
+
+        def call
+          ReturnValue.new(status: :no_action)
+        end
+      end
+    end
+  end
+end
