+++ config/initializers/omniauth.rb
@@ -4,7 +4,7 @@ Rails.application.config.middleware.use OmniAuth::Builder do
   provider :github, ENV["GITHUB_CLIENT_ID"], ENV["GITHUB_CLIENT_SECRET"],
            redirect_uri: ENV["GITHUB_APP_REDIRECT_URI"], state: SecureRandom.uuid # TODO: change state to something meaningful
 
-  provider :gitlab, ENV["GITLAB_CLIENT_ID"], ENV["GITLAB_CLIENT_SECRET"], state: SecureRandom.uuid,
+  provider :gitlab, ENV["GITLAB_APP_ID"], ENV["GITLAB_APP_SECRET"], state: SecureRandom.uuid,
            redirect_uri: ENV["GITLAB_REDIRECT_URI"]
 
   provider :slack, ENV["SLACK_CLIENT_ID"], ENV["SLACK_CLIENT_SECRET"], scope: "incoming-webhook",
